feat: 아티클 엔티티 구현
