+++ src/main/java/org/sopt/Main.java
@@ -2,8 +2,10 @@ package org.sopt;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class Main {
     public static void main(String[] args) {
         SpringApplication.run(Main.class, args);
