+++ src/main/java/org/sopt/domain/article/entity/ArticleTag.java
@@ -1,0 +1,6 @@
+package org.sopt.domain.article.entity;
+
+public enum ArticleTag {
+
+    CS, DB, SPRING, ETC
+}
