--- src/main/java/org/sopt/domain/member/entity/Article.java
@@ -1,15 +1,0 @@
-package org.sopt.domain.member.entity;
-
-import jakarta.persistence.*;
-
-@Entity
-public class Article {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "memberId")
-    private Member member;
-}
